CW #72 express-async-handler & Joi validator

+++ src/service/MflixService.mjs
@@ -33,6 +33,9 @@ export default class MflixService {
     }
     async deleteComment(id) {
         const toDeleteComment = await this.getComment(id)
+        if(!toDeleteComment) {
+            throw { code: 404, text: "comment not found" };
+        }
         await this.#commentsCollection.deleteOne({"_id":toDeleteComment._id});
         return toDeleteComment;
     }
